worker/requirements.txt
celery
redis
tenacity
psycopg2-binary>=2.9
requests